Added method: ActiveFedora::Base#required?
Created ActiveFedora::Base#required? which, when given a field name,
tells you if the field is required or not.  It figures this out by
looking for a PresenceValidator (e.g. validates_presence_of) for the
field.

This is useful so that we can create dynamically generated forms and
determine whether any given field is required or not.