Fixing delegate registry lookup
Given that the delegates hash is of indifferent access, when checking if
a delegate entry exists, we were first asking for delegates.keys. This
returned an array of strings. We were, however, passing a symbol.

By instead leveraging the hash key lookup, we don't have to worry about
receiving a symbol or a string.