Providing a YAMLAdaptor for Psych dependency
While updating from ActiveFedora 5.x to 6.4.0.rc4 I encountered a gotcha
in that Psych was now a dependency. And with Psych comes a system
dependency of libyaml. The dependency was caught on my Jenkins server
(but not Travis).

By providing the adaptor I am hoping to provide a bridge from YAML to
Psych without an abrupt requirement. As indicated in the deprecation
notice, I would envision removing the YAML adaptor.