Allow query parameter AF.reindex_everything
ActiveFedora::Base.reindex_everything certainly does that. It will
search out all objects from Fedora and attempt to load them into an
ActiveFedora::Base model. If you ran this from an application that has
only a subset of the models in Fedora you could be in for some really
weird surprises.

This patch allows you to pass a Fedora query string to the underlying
Rubydora search method.

In the case of our application, I only want to reindex items that are in
the namespace "und", so my method call looked like the following:

    `ActiveFedora::Base.reindex_everything("pid~und:*")`

Closes #160