Add caching for lookups from LDP. Also added Instrumentation and a log subscriber
If you run ActiveFedora in Rails, it injects a middleware that enables the cache
at the beginning of the request. Then if you ever ask for an object from
LDP, it caches the response and returns the cached result. At the end of
the request the cache is cleared. If you have a list of N objects and
you iterate over each to get an associated object (e.g. collection),
you'd have N loads. This prevents extra loads if the associated object
has already been loaded.