Don't skip reject_if when _destroy is passed
Nested attributes was previously skiping the check of the reject_if proc
if the _destroy flag was passed. This was occuring when allow_destroy
was false, so the record was getting created. Thanks to @ojlyytinen for
finding this bug and the orignal work on PR #879 from which this PR is
inspired.