Allow setting fcrepo_home_dir and download_dir
So that we can run `rake hydra:server` on a vm. `download_dir` allows us
to point at a location on the VM where we already have the file
downloaded.  fcrepo_home_dir enables us to point to a location that is
not a shared volume (this triggers a Fedora bug).