Update Active Triples
Few compatibility issues here that would have to be removed to support
this:

1. Order of statements that come out of property accessors is now truly
random, because of RDF.rb 2.0. A LOT of tests will have to change, and
some indexing may be unexpected.

2. There's no longer support in AT for one predicate having multiple
accessors. The logic behind that work in the past was shifty at best,
and often would result in lost data. It's for the best that it's gone -
but it's important to note.