Indexing#descendant_uris capable of prioritizing at front of list
by prioritizing URIs belonging to objects with certain hasModel literals.

By default the models used for Hydra::AccessControls permissions objects.

Because this method is used by reindex_everything, which in Hydra needs
to index permissions objects first.

The logic of keeping track of partitioned priority URLs got complex enough
that it made sense to pull it all out into it's own class, instead of trying
to add more class methods to the Indexing module.

The actual method of getting all URIs remains the same as before,
it's not fast, it requires fetching each URI from Fedora. Only change
is keeping track of fetched URIs in two sets, a priority that will be
put first in the list, and other.