more powerful ActiveFedora::FixityService
The fedora fixity service also reports the expected message
digest and file size. It makes sense to allow these to be accessed.

Actually cache the http response. Comment erroneously implied
it was cached before, but it wasn't. #check still forces a new
check, as it did before this commit, but verified? is a more
clear method name for return value, and will not bust response
cache.

Cleaned up some erroneous comments.