Allow property redefinition on `MetadataNode`
In order to change `MetadataNode`'s `:mime_type` property, we need to be able to
call `MetadataNode.property` without a `.parent_class` set. This isn't possible
normally because `.property` is overridden after the initial definition of
`:mime_type` to patch in delegation to the parent class when a new property is
defined. That delegation fails when `.parent_class` is `nil`, as is the case on
the base `MetadataNode` class.

We now only do that delegation if `.parent_class` is defined.