Rails 6 (#1403)
* Avoid private `ActiveModel::Dirty#attributes_changed_by_setter`

This method is removed in ActiveModel 6.0, and is private anyway. Avoid calling
it for upgradability.

* Drop support for Rails 5.1

Rails 5.1 support is ended. ActiveFedora 14 will drop support.

* Rely on `ActiveModel::Dirty` to track previous changes

We used to manually set `ActiveModel::Dirty` internals to track "previous
changes". In Rails 6, these internals are completely rewritten and much more
efficient. We can rely on the provided implementation instead of rolling our
own.

* Add support for Rails 6

Rails 6 is here. :tada:

* Drop support for Ruby 2.4