Porting CircleCI to 11-5 stable branch
The CircleCI config comes from the master branch, but has been modified
to reflect the configuration of the `.travis.yml` file as of the
11.5.x reality.

To address aging ruby versions, I bumped to what is EOL Ruby version
(2.3.8), but one that CircleCI may be able to build.

Further discussion:

In attempting to get this working, I encountered the case where
CircleCI fetched and installed rails 0.9.5 (see below). Not certain
why, but given that the dependencies say >= 4.2 for activerecord (and
sibling gems), I opted to set Rails development dependency.

```console
Fetching rails 0.9.5
Installing rails 0.9.5
```