improve configurability of Fedora's application path
extends the improvements in #1432 by supporting configuration of the path for
Fedora's rest endpoint. when deploying Fedora with (e.g.) Tomcat, it can be
convenient to deploy it at a path other than the server
root (e.g. `/fcrepo`). this makes it easy for applications to change this across
environments, 12-factor style.